+++ cpp/image_embed.h
@@ -6,17 +6,7 @@
 #include <tvm/runtime/container/string.h>
 #include <tvm/runtime/module.h>
 
-#ifndef MLC_LLM_DLL
-#ifdef _WIN32
-#ifdef MLC_LLM_EXPORTS
-#define MLC_LLM_DLL __declspec(dllexport)
-#else
-#define MLC_LLM_DLL __declspec(dllimport)
-#endif
-#else
-#define MLC_LLM_DLL __attribute__((visibility("default")))
-#endif
-#endif
+#include "base.h"
 
 namespace mlc {
 namespace llm {
