+++ cpp/tokenizers.cc
@@ -1,0 +1,61 @@
+/*!
+ *  Copyright (c) 2023 by Contributors
+ * \file tokenizer.cc
+ */
+
+#include "tokenizers.h"
+
+#include <tokenizers_cpp.h>
+#include <tvm/runtime/logging.h>
+
+#include <filesystem>
+#include <fstream>
+#include <string>
+
+#include "support.h"
+
+namespace mlc {
+namespace llm {
+
+std::unique_ptr<Tokenizer> TokenizerFromPath(const std::string& _path) {
+  std::filesystem::path path(_path);
+  std::filesystem::path sentencepiece;
+  std::filesystem::path huggingface;
+  std::filesystem::path rwkvworld;
+  CHECK(std::filesystem::exists(path)) << "Cannot find tokenizer via path: " << _path;
+  if (std::filesystem::is_directory(path)) {
+    sentencepiece = path / "tokenizer.model";
+    huggingface = path / "tokenizer.json";
+    rwkvworld = path / "tokenizer_model";
+    // Check ByteLevelBPE
+    {
+      std::filesystem::path merges_path = path / "merges.txt";
+      std::filesystem::path vocab_path = path / "vocab.json";
+      std::filesystem::path added_tokens_path = path / "added_tokens.json";
+      if (std::filesystem::exists(merges_path) && std::filesystem::exists(vocab_path) &&
+          std::filesystem::exists(added_tokens_path)) {
+        std::string vocab = LoadBytesFromFile(vocab_path.string());
+        std::string merges = LoadBytesFromFile(merges_path.string());
+        std::string added_tokens = LoadBytesFromFile(added_tokens_path.string());
+        return Tokenizer::FromBlobByteLevelBPE(vocab, merges, added_tokens);
+      }
+    }
+  } else {
+    sentencepiece = path.parent_path() / "tokenizer.model";
+    huggingface = path.parent_path() / "tokenizer.json";
+    rwkvworld = path.parent_path() / "tokenizer_model";
+  }
+  if (std::filesystem::exists(sentencepiece)) {
+    return Tokenizer::FromBlobSentencePiece(LoadBytesFromFile(sentencepiece.string()));
+  }
+  if (std::filesystem::exists(huggingface)) {
+    return Tokenizer::FromBlobJSON(LoadBytesFromFile(huggingface.string()));
+  }
+  if (std::filesystem::exists(rwkvworld)) {
+    return Tokenizer::FromBlobRWKVWorld(rwkvworld.string());
+  }
+  LOG(FATAL) << "Cannot find any tokenizer under: " << _path;
+}
+
+}  // namespace llm
+}  // namespace mlc
